Implemented test for page's annotations.
