+++ JDI.Light/JDI.Light.Tests/JDI.Light.Tests.csproj
@@ -73,6 +73,7 @@
     <Compile Include="Tests\Common\ComboBoxTests.cs" />
     <Compile Include="Tests\Common\DropListTests.cs" />
     <Compile Include="Tests\Composite\CustomSectionTests.cs" />
+    <Compile Include="Tests\Composite\PseudoSiteTests\AnnotationsWebPageTests.cs" />
     <Compile Include="Tests\Composite\PseudoSiteTests\CustomSectionTests.cs" />
     <Compile Include="Tests\Composite\PseudoSiteTests\CustomSectionUITests.cs" />
     <Compile Include="Tests\Composite\PseudoSiteTests\ExtendedSectionTests.cs" />
@@ -113,9 +114,13 @@
     <Compile Include="UIObjects\Pages\ComplexTablePage.cs" />
     <Compile Include="UIObjects\Pages\PerformancePage.cs" />
     <Compile Include="UIObjects\Pages\Html5Page.cs" />
-    <Compile Include="UIObjects\PseudoSections\CustomSection.cs" />
-    <Compile Include="UIObjects\PseudoSections\ExtendedSection.cs" />
-    <Compile Include="UIObjects\PseudoSite\PseudoSitePage.cs" />
+    <Compile Include="UIObjects\Pages\PseudoSite\PageWithBoth.cs" />
+    <Compile Include="UIObjects\Pages\PseudoSite\PageWithoutBoth.cs" />
+    <Compile Include="UIObjects\Pages\PseudoSite\PageWithTitle.cs" />
+    <Compile Include="UIObjects\Pages\PseudoSite\PageWithUrl.cs" />
+    <Compile Include="UIObjects\Sections\PseudoSections\CustomSection.cs" />
+    <Compile Include="UIObjects\Sections\PseudoSections\ExtendedSection.cs" />
+    <Compile Include="UIObjects\Pages\PseudoSite\PseudoSitePage.cs" />
     <Compile Include="UIObjects\Sections\Contact.cs" />
     <Compile Include="UIObjects\TestSiteBrokenDomain.cs" />
     <Compile Include="UIObjects\TestSiteCustomDomain.cs" />
