+++ JDI.Light/JDI.Light.Tests/UIObjects/Pages/PseudoSite/PageWithBoth.cs
@@ -1,0 +1,8 @@
+﻿using JDI.Light.Elements.Composite;
+
+namespace JDI.Light.Tests.UIObjects.Pages.PseudoSite
+{
+    public class PageWithBoth : WebPage
+    {
+    }
+}
